requirements.txt
asgiref==3.2.7
configparser==5.0.0
Django==3.0.6
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
mysql-connector-python==8.0.20
mysqlclient==1.3.13
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.0
PyMySQL==0.9.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
